Modularize couch db, add unit test case
