+++ db/connection.js
@@ -1,14 +1,23 @@
-const dbName = process.env.PIKSEL_DATABASE_NAME || 'piksel';
 const NodeCouchDb = require('node-couchdb');
 
-const envHost     = process.env.PIKSEL_DATABASE_HOST || '127.0.0.1';
-const envProtocol = process.env.PIKSEL_DATABASE_PROTOCOL || 'http';
-const envPort     = process.env.PIKSEL_DATABASE_PORT || '5984';
+const dbAdapter = process.env.ADAPTER || 'couch'; 
+const dbName    = process.env.PIKSEL_DATABASE_NAME || 'piksel';
 
-const couch   = new NodeCouchDb({
-    host: envHost,
-    protocol: envProtocol,
-    port: envPort
-});
+const couchEnvHost     = process.env.PIKSEL_COUCHDB_HOST || '127.0.0.1';
+const couchEnvProtocol = process.env.PIKSEL_COUCHDB_PROTOCOL || 'http';
+const couchEnvPort     = process.env.PIKSEL_COUCHDB_PORT || '5984';
 
-module.exports = { couch, dbName };
+let dbObject;
+switch(dbAdapter){
+    default:
+        dbObject = new NodeCouchDb({
+            host: couchEnvHost,
+            protocol: couchEnvProtocol,
+            port: couchEnvPort
+        });
+        break;
+}
+
+const database = dbObject;
+
+module.exports = { database, dbName };
